refactored path logic and moved some utils to jax_moseq
